submissions/Chisend/bots/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
flake8==5.0.4
idna==3.4
mccabe==0.7.0
oauthlib==3.2.1
pychimoney @ file:///Users/thelimeskies/Desktop/Projects%20/Hacktobafest/chimoney-api-community-projects/submissions/Chimoney-Python
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.10.1
urllib3==1.26.12
